+++ src/app/api/v1/results/route.ts
@@ -1,0 +1,7 @@
+import { NextResponse } from "next/server";
+import { ResultsService } from "@/features/base/results/server/service";
+
+export async function GET() {
+  const results = await ResultsService.getResultsInfo();
+  return NextResponse.json(results);
+}